+++ src/super/i-block/modules/dom/index.ts
@@ -22,7 +22,7 @@ import Friend from 'super/i-block/modules/friend';
 
 import { wait } from 'super/i-block/modules/decorators';
 import { componentRgxp } from 'super/i-block/modules/dom/const';
-import { ElCb, inViewInstanceStore } from 'super/i-block/modules/dom/interface';
+import { ElCb, inViewInstanceStore, DOMManipulationOptions } from 'super/i-block/modules/dom/interface';
 
 export * from 'super/i-block/modules/dom/const';
 export * from 'super/i-block/modules/dom/interface';
@@ -162,15 +162,19 @@ export default class DOM extends Friend {
 	 *
 	 * @param parent - element name or a link to the parent node
 	 * @param newNode
-	 * @param [group] - operation group
+	 * @param [groupOrOptions] - `async` group or a set of options
 	 */
 	appendChild(
 		parent: string | Node | DocumentFragment,
 		newNode: Node,
-		group?: string
+		groupOrOptions?: string | DOMManipulationOptions
 	): Function | false {
 		const
-			parentNode = Object.isString(parent) ? this.block?.element(parent) : parent;
+			parentNode = Object.isString(parent) ? this.block?.element(parent) : parent,
+			destroyIfComponent = Object.isPlainObject(groupOrOptions) ? groupOrOptions.destroyIfComponent : undefined;
+
+		let
+			group = Object.isString(groupOrOptions) ? groupOrOptions : groupOrOptions?.group;
 
 		if (parentNode == null) {
 			return false;
@@ -181,7 +185,18 @@ export default class DOM extends Friend {
 		}
 
 		parentNode.appendChild(newNode);
-		return this.ctx.async.worker(() => newNode.parentNode?.removeChild(newNode), {
+
+		return this.ctx.async.worker(() => {
+			newNode.parentNode?.removeChild(newNode);
+
+			const
+				{component} = <ComponentElement<iBlock>>newNode;
+
+			if (destroyIfComponent === true && component) {
+				component.unsafe.$destroy();
+			}
+
+		}, {
 			group: group ?? 'asyncComponents'
 		});
 	}
@@ -192,11 +207,15 @@ export default class DOM extends Friend {
 	 *
 	 * @param el - element name or a link to a node
 	 * @param newNode
-	 * @param [group] - operation group
+	 * @param [groupOrOptions] - `async` group or a set of options
 	 */
-	replaceWith(el: string | Element, newNode: Node, group?: string): Function | false {
+	replaceWith(el: string | Element, newNode: Node, groupOrOptions?: string | DOMManipulationOptions): Function | false {
 		const
-			node = Object.isString(el) ? this.block?.element(el) : el;
+			node = Object.isString(el) ? this.block?.element(el) : el,
+			destroyIfComponent = Object.isPlainObject(groupOrOptions) ? groupOrOptions.destroyIfComponent : undefined;
+
+		let
+			group = Object.isString(groupOrOptions) ? groupOrOptions : groupOrOptions?.group;
 
 		if (node == null) {
 			return false;
@@ -207,7 +226,18 @@ export default class DOM extends Friend {
 		}
 
 		node.replaceWith(newNode);
-		return this.ctx.async.worker(() => newNode.parentNode?.removeChild(newNode), {
+
+		return this.ctx.async.worker(() => {
+			newNode.parentNode?.removeChild(newNode);
+
+			const
+				{component} = <ComponentElement<iBlock>>newNode;
+
+			if (destroyIfComponent === true && component) {
+				component.unsafe.$destroy();
+			}
+
+		}, {
 			group: group ?? 'asyncComponents'
 		});
 	}