+++ src/super/i-block/modules/dom/CHANGELOG.md
@@ -9,11 +9,13 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
-## v3.0.0-rc.60 (2020-08-??)
+## v3.0.0-rc.60 (2020-09-01)
 
 #### :rocket: New Feature
 
 * [Added `watchForIntersection` method and `localInView` getter](https://github.com/V4Fire/Client/issues/195)
+* [Added an option `destroyIfComponent` into `i-block/dom/replaceWith`, `i-block/dom/appendChild`](https://github.com/V4Fire/Client/pull/321)
+
 
 ## v3.0.0-rc.46 (2020-07-31)